+++ src/components/address-form/address-form.module.css
@@ -1,0 +1,20 @@
+.form {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 2rem;
+}
+
+.details {
+  row-gap: 1rem;
+}
+
+.formItem {
+  flex: 0 0 calc(50% - 1rem);
+  max-width: 100%;
+}
+
+@media (width <= 768px) {
+  .formItem {
+    flex: 0 0 100%;
+  }
+}
