--- src/components/address-form/address.module.css
@@ -1,33 +1,0 @@
-.defaultLabel {
-  font-weight: 500;
-  color: var(--color-accent);
-}
-
-.fieldName {
-  margin-right: 1rem;
-  color: var(--color-accent-light);
-}
-
-.form,
-.details {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 2rem;
-}
-
-.details {
-  row-gap: 1rem;
-}
-
-.formItem,
-.detailsItem {
-  flex: 0 0 calc(50% - 1rem);
-  max-width: 100%;
-}
-
-@media (width <= 768px) {
-  .formItem,
-  .detailsItem {
-    flex: 0 0 100%;
-  }
-}
